requirements.txt
tensorflow==2.0.0a0
Werkzeug==0.14.1
Flask==1.0.2
numpy==1.16.2
Pillow==7.0.0
gunicorn>=19.9.0
